backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
PyPDF2==3.0.1
python-dotenv==0.19.0
requests==2.26.0
langchain==0.0.267
openai==0.27.0
werkzeug==2.0.3 